refactor warn-once into SRP submodules
